// src/api/generated/models/collectionResponse.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Dolink API Documentation
 * This is DoLink! 저장하고, 묶고, 바로 꺼내는 링크 관리
 * OpenAPI spec version: v1.0.0
 */
import type { CollectionResponseCategory } from './collectionResponseCategory';

export interface CollectionResponse {
  collectionId?: number;
  name?: string;
  category?: CollectionResponseCategory;
  thumbnails?: string[];
  taskCount?: number;
  isTutorial?: boolean;
}